pyproject.toml
[tool.poetry]
name = "fiability"
version = "0.1.0"
description = "Fiability check tool"
authors = ["FIAbility"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9,<3.11"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
peewee = "^3.14.8"
PyMySQL = "^1.0.2"
newspaper3k = "^0.2.8"
scikit-learn = "^1.0.2"
nltk = "^3.6.7"
urllib3 = "^1.26.8"
pytest = "^6.2.5"
validators = "^0.18.2"
pandas = "^1.4.0"
python-dotenv = "^0.19.2"
gunicorn = "^20.1.0"
requests = "^2.27.1"
matplotlib = "^3.5.1"
seaborn = "^0.11.2"
termcolor = "^1.1.0"
langdetect = "^1.0.9"
gensim = "^4.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
